octomap_server_launch: gompertz expiry parameterization
The octomap server was changed to use a gompertz curve for expiration
times instead of a simple quadratic. This allows more control over the
expiration when a voxel has more observations and prevents premature
"burn-in" by making the curve a sigmoid instead of a parabola.
Re-paremterize all the octomap servers with the new gompertz curve by
mostly matching the parabola in the uncertain region of observation
probablilty hits, but taper off towards the maximum expiration instead
of dramatically as the quadratic did.