front lidar: move to raw_z_filtered
For clearing-only cloud use raw_z_filtered to only clear invalid that
is not under the floor. Under-floor invalid is very likely to be
reflections. Also, the raw_z_filtered is above the height used for
the static map, so this will keep clearing through invalid from
decimating the static map layer for global 3D costmaps.