octomap_server_launch: set skip_count instead of throttled topics
Keeping throttled topics in sync is not something the existing throttle
nodelets do. They use now in each instance for comparing times. Even if
they were switched to the header stamp it is still possible to get out
of sync at the very start (as the nodelets load at differen times) or if
any glitch caused a message to be lost by some but not all parts of the
driver stacks (they do use queue sizes of one). Work around this by no
longer using throttled topics, but using the skip_count feature in the
octomap server instead. This still saves most of the processing time.