octomap_server_launch: fully saltify params and service
Fully saltify the params and service (occupancy-mapping-3d). Move the
parameterization from the old static yaml files in the params directory
to the salt pillar. Salt-enable the occupancy-mapping-3d service so it
will restart on any parameter change.

As part of this change, remove the redundant '_map' from namespace.
This is possible now that the parameters are in salt. Previously the
static params files needed different names, and the sensor_name arg was
being used for that. These changes need to be combined as they make use
of each other (due to the pillar layout).