Remove default value for config_dir salt key
Rather than specifying default values for the badger configuration
directory, let's just require that the salt key and environment variable
exist.

Changing the value of the configuration directory becomes a maintenance
nightmare if we have 100s of default values scattered throughout the
code base that we have to go update.