Add missing default values for salt variables
When the salt setting changes are made to enforce that keys exist or a
that a default value is specified this code could throw a KeyError if
the expected data isn't in the pillar.

Note: The `configuration:hardware:geometry:clearance` salt variable is
required to exist so don't add a default and don't check with an "if"
later in the file.